+++ public/.babelrc
@@ -1,0 +1,5 @@
+{
+    "plugins": [
+        "@babel/plugin-proposal-class-properties"
+    ]
+}
