+++ public/index.html
@@ -18,16 +18,15 @@
 		<div class="for_goods">
 			<div class="goods-list"></div>
 		</div>
-		<div class="get_expression">
-
+		<!--<div class="get_expression">
 			<a href="./lesson4.html" class="btn_lesson4">Урок 4. Нажмите, чтобы перейти на новую страницу</a>
-
 		</div>
-	</div>
+	</div>-->
 
-	<script src="./js/button.js"></script>
-	<script src="./js/shop.js"></script>
-	<script src="./js/lesson4.js"></script>
+		<!--script src="./js/button.js"></script>
+	//<script src="./js/shop.js"></script>
+	//<script src="./js/lesson4.js"></script> -->
+		<script src="./js/shop.js"></script>
 
 
 </body>
