+++ public/js/main.js
@@ -1,0 +1,1 @@
+(()=>{"use strict";class t{constructor(t,e){this._text=t,this.callback=e}onBtnClick(){const t=this.callback;"function"==typeof t&&t()}getTempLate(){const t=document.createElement("button");return t.classList.add("btn"),t}render(t){if(t){const e=this.getTempLate();e.innerHTML=this._text,t.appendChild(e),e.addEventListener("click",(()=>{this.onBtnClick()}))}}}class e{constructor({name:t,price:e,img:n}){this._name=t,this._price=e,this._img=n}addToCart(){const t=document.querySelector(".basket");if(t){t.innerHTML="";const e=document.createElement("div");e.innerHTML=`Tовар "${this._name}" в количестве ${this.counter++} шт.           \n            добавлен в корзину </BR> </BR>`,t.appendChild(e);const n=document.createElement("div");s++,i+=this._price,n.innerHTML="Всего товаров в корзине: "+s+" шт. на сумму "+i+" руб.",t.appendChild(n)}}render(){const e=document.querySelector(".goods-list");if(e){const n=document.createElement("div");n.innerHTML=`${this._name} = ${this._price} руб. </BR>\n            <img src = "${this._img}" /> </BR>`,new t("Добавить в корзину",this.addToCart.bind(this)).render(n),e.appendChild(n)}}}const n=document.querySelector(".btn_basket");if(n){const e=document.createElement("div");e.innerHTML="Корзина",n.appendChild(e),new t("Корзина",n)}let s=0,i=0;new class{constructor(t){this.fetchGoods().then((t=>t.json())).then((n=>{const s=n.data.map((n=>new e(n,t)));this._items=s,this.render()}))}fetchGoods(){return fetch("http://localhost:3000/database/items.json")}render(){this._items.forEach((t=>{t.render()}))}}})();
