+++ public/js/shop.js
@@ -1,90 +1,1 @@
-class List {
-    _items = []
-    constructor(CartInstane) {
-        this.fetchGoods()
-            .then(res => {
-                return res.json()
-            })
-            .then(data => {
-                const goods = data.data.map(item => {
-                    return new GoodItem(item, CartInstane)
-                })
-                this._items = goods
-                this.render()
-            })
-    }
-
-    fetchGoods() {
-        const url = 'http://localhost:3000/database/items.json'
-        return fetch(url);
-    }
-
-    render() {
-        this._items.forEach(Good => {
-            Good.render()
-        })
-    }
-
-}
-
-class GoodItem {
-    _name = ''
-    _price = 0
-    _img = 0
-    counter = 1
-    _allCounter = 1
-
-
-    constructor({
-        name,
-        price,
-        img
-    }) {
-        this._name = name
-        this._price = price
-        this._img = img
-    }
-
-
-    addToCart() {
-        const placeToRender = document.querySelector('.basket')
-        if (placeToRender) {
-            placeToRender.innerHTML = ''
-            const addCart = document.createElement('div')
-            addCart.innerHTML = `Tовар "${this._name}" в количестве ${this.counter++} шт.           
-            добавлен в корзину </BR> </BR>`
-            placeToRender.appendChild(addCart)
-            const allCart = document.createElement('div')
-            quantityGoods++
-            allPriceGoods = allPriceGoods + this._price
-            allCart.innerHTML = "Всего товаров в корзине: " + quantityGoods + " шт. на сумму " + allPriceGoods + " руб."
-            placeToRender.appendChild(allCart)
-        }
-
-    }
-
-    render() {
-        const placeToRender = document.querySelector('.goods-list')
-        if (placeToRender) {
-            const block = document.createElement('div')
-            block.innerHTML = `${this._name} = ${this._price} руб. </BR>
-            <img src = "${this._img}" /> </BR>`
-            const btn = new Button('Добавить в корзину', this.addToCart.bind(this))
-            btn.render(block)
-            placeToRender.appendChild(block)
-        }
-    }
-}
-
-
-const placeToCart = document.querySelector('.btn_basket')
-if (placeToCart) {
-    const btnCart = document.createElement('div')
-    btnCart.innerHTML = `Корзина`
-    placeToCart.appendChild(btnCart)
-    new Button('Корзина', placeToCart)
-}
-
-let quantityGoods = 0,
-    allPriceGoods = 0
-new List()
+(()=>{"use strict";var n={112:n=>{n.exports=function(n){var e=[];return e.toString=function(){return this.map((function(e){var t=n(e);return e[2]?"@media ".concat(e[2]," {").concat(t,"}"):t})).join("")},e.i=function(n,t,r){"string"==typeof n&&(n=[[null,n,""]]);var i={};if(r)for(var o=0;o<this.length;o++){var a=this[o][0];null!=a&&(i[a]=!0)}for(var s=0;s<n.length;s++){var c=[].concat(n[s]);r&&i[c[0]]||(t&&(c[2]?c[2]="".concat(t," and ").concat(c[2]):c[2]=t),e.push(c))}},e}},426:(n,e,t)=>{t.d(e,{Z:()=>o});var r=t(112),i=t.n(r)()((function(n){return n[1]}));i.push([n.id,"* {\r\n    margin: 0;\r\n    padding: 0;\r\n}\r\n\r\n.proect {\r\n    width: 1400px;\r\n    height: 1400px;\r\n    background-color: rgb(101, 111, 121);\r\n}\r\n\r\n.for_goods {\r\n    width: 1200px;\r\n    padding-top: 20px;\r\n    padding-left: 100px;\r\n    padding-right: 100px;\r\n    font-size: 22px;\r\n    color: white;\r\n}\r\n\r\n.goods-list {\r\n    height: 700px;\r\n    display: flex;\r\n    flex-wrap: wrap;\r\n    justify-content: space-between;\r\n}\r\n\r\n.btns {\r\n    padding-top: 20px;\r\n    height: 120px;\r\n    padding-left: 100px;\r\n    padding-right: 100px;\r\n\r\n    background-color: rgb(101, 111, 121);\r\n\r\n    color: beige;\r\n    border: 3px solid #fff;\r\n    /* Белая рамка */\r\n    border-radius: 50px;\r\n    /* Радиус скругления */\r\n    display: flex;\r\n    flex-wrap: wrap;\r\n    justify-content: space-between;\r\n}\r\n\r\n.basket {\r\n    width: 900px;\r\n    height: 100px;\r\n    font-size: 25px;\r\n}\r\n\r\n.btn_basket {\r\n    font-size: 55px;\r\n    width: 200px;\r\n}\r\n\r\n.get_expression {\r\n    padding-left: 100px;\r\n    padding-top: 20px;\r\n    height: 200px;\r\n\r\n}\r\n\r\n.goods-search {\r\n    font-size: 25px;\r\n}\r\n\r\n.btn_lesson4 {\r\n    padding-left: 20px;\r\n    padding-right: 20px;\r\n    font-size: 35px;\r\n    color: black;\r\n    background-color: white;\r\n    border-radius: 50px;\r\n    text-decoration: none;\r\n}\r\n\r\n.regist {\r\n    padding-left: 20px;\r\n    font-size: 25px;\r\n    color: white;\r\n}\r\n\r\n.regist_input {\r\n    font-size: 25px;\r\n}\r\n\r\n.input_bloc {\r\n    margin-bottom: 10px;\r\n    display: flex;\r\n    font-size: 25px;\r\n}",""]);const o=i},379:(n,e,t)=>{var r,i=function(){var n={};return function(e){if(void 0===n[e]){var t=document.querySelector(e);if(window.HTMLIFrameElement&&t instanceof window.HTMLIFrameElement)try{t=t.contentDocument.head}catch(n){t=null}n[e]=t}return n[e]}}(),o=[];function a(n){for(var e=-1,t=0;t<o.length;t++)if(o[t].identifier===n){e=t;break}return e}function s(n,e){for(var t={},r=[],i=0;i<n.length;i++){var s=n[i],c=e.base?s[0]+e.base:s[0],d=t[c]||0,l="".concat(c," ").concat(d);t[c]=d+1;var p=a(l),u={css:s[1],media:s[2],sourceMap:s[3]};-1!==p?(o[p].references++,o[p].updater(u)):o.push({identifier:l,updater:m(u,e),references:1}),r.push(l)}return r}function c(n){var e=document.createElement("style"),r=n.attributes||{};if(void 0===r.nonce){var o=t.nc;o&&(r.nonce=o)}if(Object.keys(r).forEach((function(n){e.setAttribute(n,r[n])})),"function"==typeof n.insert)n.insert(e);else{var a=i(n.insert||"head");if(!a)throw new Error("Couldn't find a style target. This probably means that the value for the 'insert' parameter is invalid.");a.appendChild(e)}return e}var d,l=(d=[],function(n,e){return d[n]=e,d.filter(Boolean).join("\n")});function p(n,e,t,r){var i=t?"":r.media?"@media ".concat(r.media," {").concat(r.css,"}"):r.css;if(n.styleSheet)n.styleSheet.cssText=l(e,i);else{var o=document.createTextNode(i),a=n.childNodes;a[e]&&n.removeChild(a[e]),a.length?n.insertBefore(o,a[e]):n.appendChild(o)}}function u(n,e,t){var r=t.css,i=t.media,o=t.sourceMap;if(i?n.setAttribute("media",i):n.removeAttribute("media"),o&&"undefined"!=typeof btoa&&(r+="\n/*# sourceMappingURL=data:application/json;base64,".concat(btoa(unescape(encodeURIComponent(JSON.stringify(o))))," */")),n.styleSheet)n.styleSheet.cssText=r;else{for(;n.firstChild;)n.removeChild(n.firstChild);n.appendChild(document.createTextNode(r))}}var f=null,h=0;function m(n,e){var t,r,i;if(e.singleton){var o=h++;t=f||(f=c(e)),r=p.bind(null,t,o,!1),i=p.bind(null,t,o,!0)}else t=c(e),r=u.bind(null,t,e),i=function(){!function(n){if(null===n.parentNode)return!1;n.parentNode.removeChild(n)}(t)};return r(n),function(e){if(e){if(e.css===n.css&&e.media===n.media&&e.sourceMap===n.sourceMap)return;r(n=e)}else i()}}n.exports=function(n,e){(e=e||{}).singleton||"boolean"==typeof e.singleton||(e.singleton=(void 0===r&&(r=Boolean(window&&document&&document.all&&!window.atob)),r));var t=s(n=n||[],e);return function(n){if(n=n||[],"[object Array]"===Object.prototype.toString.call(n)){for(var r=0;r<t.length;r++){var i=a(t[r]);o[i].references--}for(var c=s(n,e),d=0;d<t.length;d++){var l=a(t[d]);0===o[l].references&&(o[l].updater(),o.splice(l,1))}t=c}}}}},e={};function t(r){if(e[r])return e[r].exports;var i=e[r]={id:r,exports:{}};return n[r](i,i.exports,t),i.exports}t.n=n=>{var e=n&&n.__esModule?()=>n.default:()=>n;return t.d(e,{a:e}),e},t.d=(n,e)=>{for(var r in e)t.o(e,r)&&!t.o(n,r)&&Object.defineProperty(n,r,{enumerable:!0,get:e[r]})},t.o=(n,e)=>Object.prototype.hasOwnProperty.call(n,e),(()=>{function n(n,e,t){return e in n?Object.defineProperty(n,e,{value:t,enumerable:!0,configurable:!0,writable:!0}):n[e]=t,n}class e{constructor(e,t){n(this,"_text",""),n(this,"_callback",null),this._text=e,this.callback=t}onBtnClick(){const n=this.callback;"function"==typeof n&&n()}getTempLate(){const n=document.createElement("button");return n.classList.add("btn"),n}render(n){if(n){const e=this.getTempLate();e.innerHTML=this._text,n.appendChild(e),e.addEventListener("click",(()=>{this.onBtnClick()}))}}}var r=t(379),i=t.n(r),o=t(426);function a(n,e,t){return e in n?Object.defineProperty(n,e,{value:t,enumerable:!0,configurable:!0,writable:!0}):n[e]=t,n}i()(o.Z,{insert:"head",singleton:!1}),o.Z.locals;class s{constructor({name:n,price:e,img:t}){a(this,"_name",""),a(this,"_price",0),a(this,"_img",0),a(this,"counter",1),a(this,"_allCounter",1),this._name=n,this._price=e,this._img=t}addToCart(){const n=document.querySelector(".basket");if(n){n.innerHTML="";const e=document.createElement("div");e.innerHTML=`Tовар "${this._name}" в количестве ${this.counter++} шт.           \n            добавлен в корзину </BR> </BR>`,n.appendChild(e);const t=document.createElement("div");d++,l+=this._price,t.innerHTML="Всего товаров в корзине: "+d+" шт. на сумму "+l+" руб.",n.appendChild(t)}}render(){const n=document.querySelector(".goods-list");if(n){const t=document.createElement("div");t.innerHTML=`${this._name} = ${this._price} руб. </BR>\n            <img src = "${this._img}" /> </BR>`,new e("Добавить в корзину",this.addToCart.bind(this)).render(t),n.appendChild(t)}}}const c=document.querySelector(".btn_basket");if(c){const n=document.createElement("div");n.innerHTML="Корзина",c.appendChild(n),new e("Корзина",c)}let d=0,l=0;new class{constructor(n){a(this,"_items",[]),this.fetchGoods().then((n=>n.json())).then((e=>{const t=e.data.map((e=>new s(e,n)));this._items=t,this.render()}))}fetchGoods(){return fetch("http://127.0.0.1:5500/public/database/items.json")}render(){this._items.forEach((n=>{n.render()}))}}})()})();
