+++ public/package.json
@@ -4,8 +4,9 @@
   "description": "nadin webserver",
   "main": "script.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node script.js"
+    "test": "echo \"Error: no test specified\" && exit 1",    
+    "build": "webpack",    
+    "dev": "webpack serve --hot --config webpack.dev.config.js"
   },
   "repository": {
     "type": "git",
@@ -16,5 +17,16 @@
   "bugs": {
     "url": "https://github.com/NADIN-git/Advanced_JS_Course/issues"
   },
-  "homepage": "https://github.com/NADIN-git/Advanced_JS_Course#readme"
+  "homepage": "https://github.com/NADIN-git/Advanced_JS_Course#readme",
+  "devDependencies": {
+    "webpack": "^5.24.3",
+    "webpack-cli": "^4.5.0",
+    "webpack-merge": "^5.7.3"
+  },
+  "dependencies": {
+    "@babel/core": "^7.13.8",
+    "css-loader": "^5.1.1",
+    "style-loader": "^2.0.0",
+    "webpack-dev-server": "^3.11.2"
+  }
 }
