+++ public/src/style.css
@@ -1,0 +1,92 @@
+* {
+    margin: 0;
+    padding: 0;
+}
+
+.proect {
+    width: 1400px;
+    height: 1400px;
+    background-color: rgb(101, 111, 121);
+}
+
+.for_goods {
+    width: 1200px;
+    padding-top: 20px;
+    padding-left: 100px;
+    padding-right: 100px;
+    font-size: 22px;
+    color: white;
+}
+
+.goods-list {
+    height: 700px;
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-between;
+}
+
+.btns {
+    padding-top: 20px;
+    height: 120px;
+    padding-left: 100px;
+    padding-right: 100px;
+
+    background-color: rgb(101, 111, 121);
+
+    color: beige;
+    border: 3px solid #fff;
+    /* Белая рамка */
+    border-radius: 50px;
+    /* Радиус скругления */
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-between;
+}
+
+.basket {
+    width: 900px;
+    height: 100px;
+    font-size: 25px;
+}
+
+.btn_basket {
+    font-size: 55px;
+    width: 200px;
+}
+
+.get_expression {
+    padding-left: 100px;
+    padding-top: 20px;
+    height: 200px;
+
+}
+
+.goods-search {
+    font-size: 25px;
+}
+
+.btn_lesson4 {
+    padding-left: 20px;
+    padding-right: 20px;
+    font-size: 35px;
+    color: black;
+    background-color: white;
+    border-radius: 50px;
+    text-decoration: none;
+}
+
+.regist {
+    padding-left: 20px;
+    font-size: 25px;
+    color: white;
+}
+
+.regist_input {
+    font-size: 25px;
+}
+
+.input_bloc {
+    margin-bottom: 10px;
+    display: flex;
+    font-size: 25px;
+}
