+++ public/webpack.dev.config.js
@@ -1,0 +1,14 @@
+const {
+    merge
+} = require('webpack-merge')
+const base = require('./webpack.config')
+
+module.exports = merge(base, {
+    devServer: {
+        contentBase: './database',
+        publicPath: '/.js',
+        host: 'localhost',
+        port: 3000,
+        hot: true
+    }
+})
